Move page.Limiter to actor.Limiter
