+++ internal/page/donor/witnessing_as_independent_witness.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/ministryofjustice/opg-go-common/template"
+	"github.com/ministryofjustice/opg-modernising-lpa/internal/actor"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/page"
 	"github.com/ministryofjustice/opg-modernising-lpa/internal/validation"
 )
@@ -29,7 +30,7 @@ func WitnessingAsIndependentWitness(tmpl template.Template, donorStore DonorStor
 			data.Errors = data.Form.Validate()
 
 			if lpa.WitnessCodeLimiter == nil {
-				lpa.WitnessCodeLimiter = page.NewLimiter(time.Minute, 5, 10)
+				lpa.WitnessCodeLimiter = actor.NewLimiter(time.Minute, 5, 10)
 			}
 
 			if !lpa.WitnessCodeLimiter.Allow(now()) {
